import 2.5 enhancement #11380 port ORCHdfsFileStream to OrcScanner to 2.3
